HOTFIX MEMORY LEAK puresample_training loop
Keras model integer constants were being used to maintain epoch for the
model. At the end of the training loop these were being incremented by a
method. Doing this increment using python resulted in the entire
computational graph for TF being redrawn for the following training
steps. Need to look into this. It is possible the keras model checks for
if any internal values have changed and redraws if true. To get around
this problem The cooling distance (epoch) is now defined as a tensorflow
constant and is incrememnted by an out of graph eager function. Initial
tests show it works